Refactor customer name input and account lookup
Added an initialize method to BankData because a BankData object with
no data is not very useful. The initialize method requires that a
customer name be passed into it. Also cleaned up the #isvalid? method
and removed output because it doesn't belong there.

Teller class requires a #customer_valid? method which calls #is_valid?
on BankData object. Also renamed tell_prompt to menu to make things
easier to read.

The menu can now ask the customer for their name. Also added a reprimand
customer method that tells the customer that the name they entered is
invalid.

Moved the validation loop from teller to atm. Also refactored the
validation loop.