Pushed deposit behavior into BankData
The Teller class had too many responsibilities. The Teller had access
to all of the customers information, which it does not need. By pushing
the deposit making responsibilities to the BankData class, the
 #make_deposit method is easier to understand and also looks much
cleaner.

The "balance" instance variable from BankData has been removed, and
a method was added that returns the balance for the customer name. This
way there are less instances of the customer's balance to update and
keep up with.

At this point the #make_withdrawal method does not work. It still uses
the balance instance variable that was deleted, and is also still very
messy.