Refactored withdraw methods and output
Changed the withdraw method so that the dispenser only does one
calculation based off of the withdraw amount * 100. This also makes the
bill and coin constants one conjoined variable. The code is much
shorter and makes much more sense

Changed the withdraw output so that the 'amounts' hash is the only
thing that needs to be passed in as a parameter. The constants were
moved to the menu class and accessed directly, and the hash is now only
one hash instead of one for bills and coins separately.

Also changed the name of 'reprimand_customer' to 'invalid_customer',
because we really shouldn't be reprimanding our customers.