Bankdata and Teller class refactor
Moved menu operations outside of the BankData class. This way the
menu doesn't have to be passed all the way down into the BankData
methods.

Moved the name getting and setting methods into the teller class,
and made an instance variable to keep up with the name for later use.

The deposit_funds and withdraw_funds BankData methods now return
the new_balance instead of storing in a local variable.