Changed customer creation and validation logic
Added an instance variable to Teller class to keep up with the
account balance. Changed the way that the Teller instantiates a
new customer.

Changed the logic in atm.rb so that an invalid customer name outputs
the invalid_customer menu output. The logic is much easier to read and
understand.

Changed Teller and BankData instance variables to attr_reader instead
of attr_accesor, since all they need to do is be readable.

Changed set_name to customer_setup inside the Teller class, and made the
method set the BankData object as well as the balance instance variable.