Completely refactored Teller class and atm.rb
Created a work method inside the Teller class that executes the jobs of
a teller based on the Menu and BankData objects. Both objects are now
passed to the Teller instantiation.

Created a get_request method inside the Teller class. This method
returns an array containing the request keyword, and the amount of
withdraw/deposit (if needed). Deleted the customer_valid method inside
the Teller class, and used the balance instance variable instead.

atm.rb is much shorter. It just creates the Menu and BankData objects,
passes them into the Teller instantiation, and tells the Teller to get
to work.

Changed some of the puts statements in the Menu class to work better
with the new Teller logic, and to make the screen output more readable.