Made changes to CashDispenser, Teller, and Menu
Changed CashDispenser so that the dispense method returns a hash
containing the denomination as a string (fives, tens, ones, etc.),
and the quantity of the denomination. Now the dispenser method no
longer has to be passed any constants, and can return only what
the menu needs to display the results.

Changed several if blocks in the Menu to a single line. Changed the
main menu to ask for a number instead of the request as a string.
Combined all Menu methods into two methods: A prompt method that returns
customer input, and a display method that only show a message to the
customer.

Changed logic inside Teller methods. Created a REQUEST_OPTIONS constant
that is used to change the number returned from the main menu query to
the appropriate string keyword. The request keyword is used to call
another Teller method using the .send method, or as an argument to a
call to the display method of the Menu class.